added graph input

+++ src/weightedGraph/Edge.java
@@ -9,16 +9,16 @@ public class Edge
 {
     private final int v;
     private final int w;
-    private final double weight;
+    private final int weight;
 
-    public Edge(int v, int w, double weight)
+    public Edge(int v, int w, int weight)
     {
         this.v = v;
         this.w = w;
         this.weight = weight;
     }
 
-    public double weight()
+    public int weight()
     {
         return weight;
     }
